refactor : port를 query, command로 패키지 분류
