+++ src/main/java/com/yapp/artie/domain/user/adapter/in/web/GetTestUserTokenController.java
@@ -1,7 +1,7 @@
 package com.yapp.artie.domain.user.adapter.in.web;
 
-import com.yapp.artie.domain.user.application.port.in.GetTestUserTokenQuery;
-import com.yapp.artie.domain.user.application.port.in.GetTestUserTokenResponse;
+import com.yapp.artie.domain.user.application.port.in.query.GetTestUserTokenQuery;
+import com.yapp.artie.domain.user.application.port.in.response.GetTestUserTokenResponse;
 import com.yapp.artie.global.common.annotation.WebAdapter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
