+++ src/main/java/com/yapp/artie/domain/user/adapter/in/web/GetUserController.java
@@ -1,6 +1,6 @@
 package com.yapp.artie.domain.user.adapter.in.web;
 
-import com.yapp.artie.domain.user.application.port.in.GetUserQuery;
+import com.yapp.artie.domain.user.application.port.in.query.GetUserQuery;
 import com.yapp.artie.domain.user.domain.User;
 import com.yapp.artie.global.common.annotation.WebAdapter;
 import io.swagger.v3.oas.annotations.Operation;
