+++ src/main/java/com/yapp/artie/domain/user/adapter/in/web/GetUserThumbnailController.java
@@ -1,8 +1,8 @@
 package com.yapp.artie.domain.user.adapter.in.web;
 
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
-import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailQuery;
-import com.yapp.artie.domain.user.application.port.in.GetUserThumbnailResponse;
+import com.yapp.artie.domain.user.application.port.in.query.GetUserThumbnailQuery;
+import com.yapp.artie.domain.user.application.port.in.response.GetUserThumbnailResponse;
 import com.yapp.artie.global.common.annotation.WebAdapter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
