--- src/main/java/com/yapp/artie/domain/user/application/port/in/GetUserThumbnailQuery.java
@@ -1,6 +1,0 @@
-package com.yapp.artie.domain.user.application.port.in;
-
-public interface GetUserThumbnailQuery {
-
-  GetUserThumbnailResponse loadUserThumbnailById(Long id);
-}
