--- src/main/java/com/yapp/artie/domain/user/application/port/in/RegisterUserUseCase.java
@@ -1,6 +1,0 @@
-package com.yapp.artie.domain.user.application.port.in;
-
-public interface RegisterUserUseCase {
-
-  RegisterUserResponse register(String uid, String username, String picture);
-}
