+++ src/main/java/com/yapp/artie/domain/user/application/service/GetUserDetailsService.java
@@ -2,7 +2,7 @@ package com.yapp.artie.domain.user.application.service;
 
 import static org.springframework.security.core.userdetails.User.builder;
 
-import com.yapp.artie.domain.user.application.port.in.GetUserDetailsQuery;
+import com.yapp.artie.domain.user.application.port.in.query.GetUserDetailsQuery;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
 import com.yapp.artie.global.common.annotation.UseCase;
