+++ src/main/java/com/yapp/artie/domain/user/application/service/GetUserService.java
@@ -1,6 +1,6 @@
 package com.yapp.artie.domain.user.application.service;
 
-import com.yapp.artie.domain.user.application.port.in.GetUserQuery;
+import com.yapp.artie.domain.user.application.port.in.query.GetUserQuery;
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
 import com.yapp.artie.global.common.annotation.UseCase;
